# booking-api-backend/booking/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TicketFlow API",
    "version": "1.0.0",
    "description": "API documentation for TicketFlow"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check endpoint",
        "tags": ["System"],
        "security": [],
        "responses": {
          "200": {
            "description": "System is healthy"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Registration successful" },
          "409": { "description": "Email already exists" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get current user profile",
        "tags": ["Auth"],
        "responses": {
          "200": { "description": "User profile" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/categories": {
      "get": {
        "summary": "List all categories",
        "tags": ["Categories"],
        "security": [],
        "responses": {
          "200": { "description": "List of categories" }
        }
      },
      "post": {
        "summary": "Create a new category (Admin only)",
        "tags": ["Categories"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Category created" }
        }
      }
    }
  }
}